+++ firebase_rules_generator/test/database/match.dart
@@ -1,5 +1,5 @@
 import 'util.dart';
 
 void main() {
-  testDatabaseRulesBuilder('test');
+  testDatabaseRulesBuilder('match');
 }