+++ firebase_rules_generator/test_project/lib/firebase/raw/input.dart
@@ -1,0 +1,11 @@
+// ignore_for_file: unused_local_variable
+
+import 'package:firebase_rules/firebase.dart';
+
+bool test() {
+  final a = rules.parseBool('true'.rules);
+  return true;
+}
+
+@FirebaseRules(service: Service.firestore, functions: [test])
+final firestoreRules = <Match>[];