+++ firebase_rules_generator/lib/src/firebase/sanitizer.dart
@@ -5,79 +5,92 @@ import 'package:firebase_rules_generator/src/firebase/revived_firebase_rules.dar
 
 /// Sanitize rules files
 String sanitizeRules(RevivedFirebaseRules annotation, String input) {
-  final pass1 = removeRulesPrefixesAndSuffixes(input);
-  final pass2 = stripNullSafety(pass1);
-  final pass3 = translateStrings(pass2);
+  return transformIgnoringRaws(input, [
+    removeRulesPrefixesAndSuffixes,
+    stripNullSafety,
+    translateStrings,
+    (input) => input
+            // Convert firestore methods
+            .replaceAllMapped(
+          RegExp(r'firestore\.(.+?)(<.+?>)?\((.+?)\)'),
+          (m) {
+            final buffer = StringBuffer();
+            if (annotation.service != Service.firestore) {
+              buffer.write('firestore.');
+            }
+            buffer.write('${m[1]}(${m[3]})');
+            return buffer.toString();
+          },
+        )
+            // Convert all paths
+            .replaceAllMapped(
+          RegExp(r"path\('(.+?)'(, database: '(.+?)')?\)"),
+          (m) {
+            final database = m[3];
+            final String path;
+            if (database == null) {
+              path = '/databases/\$(database)/documents${m[1]}';
+            } else {
+              path = '/databases/($database)/documents${m[1]}';
+            }
+            return 'path(\'$path\')';
+          },
+        ),
+    (input) => input
+        // Convert `contains` to `x in y`
+        .replaceAllMapped(
+          RegExp(r'(\S+?|\[.+?|\{.+?)\.contains\((.+?)\)'),
+          (m) => '${m[2]} in ${m[1]}',
+        )
+        // Convert `range` to `x[i:j]
+        .replaceAllMapped(
+          RegExp(r'(\S+?)\.range\((.+?), (.+?)\)'),
+          (m) => '${m[1]}[${m[2]}:${m[3]}]',
+        ),
+    (input) => input
+        // bool parsing
+        .replaceAllMapped(RegExp(r'parseBool\((.+?)\)'), (m) => 'bool(${m[1]})')
+        // bytes parsing
+        .replaceAllMapped(RegExp(r"parseBytes\('(.+?)'\)"), (m) => "b'${m[1]}'")
+        // float parsing
+        .replaceAllMapped(
+            RegExp(r'parseFloat\((.+?)\)'), (m) => 'float(${m[1]})')
+        // int parsing
+        .replaceAllMapped(RegExp(r'parseInt\((.+?)\)'), (m) => 'int(${m[1]})'),
+    (input) => translateEnums(input, {
+          'RulesDurationUnit': RulesDurationUnit.values,
+          'RulesMethod': RulesMethod.values,
+          'RulesIdentityProvider': RulesIdentityProvider.values,
+          'RulesSignInProvider': RulesSignInProvider.values,
+        }),
+    translateAuthVariables,
+    (input) => input.replaceAll(
+          'resource.firestoreResourceName',
+          "resource['__name__']",
+        ),
+  ]);
+}
 
-  final pass4 = pass3
-      // Convert firestore methods
-      .replaceAllMapped(
-    RegExp(r'firestore\.(.+?)(<.+?>)?\((.+?)\)'),
-    (m) {
-      final buffer = StringBuffer();
-      if (annotation.service != Service.firestore) {
-        buffer.write('firestore.');
-      }
-      buffer.write('${m[1]}(${m[3]})');
-      return buffer.toString();
-    },
-  )
-      // Convert all paths
-      .replaceAllMapped(
-    RegExp(r"path\('(.+?)'(, database: '(.+?)')?\)"),
-    (m) {
-      final database = m[3];
-      final String path;
-      if (database == null) {
-        path = '/databases/\$(database)/documents${m[1]}';
-      } else {
-        path = '/databases/($database)/documents${m[1]}';
-      }
-      return 'path(\'$path\')';
-    },
-  );
-
-  final pass5 = pass4
-      // Convert `contains` to `x in y`
-      .replaceAllMapped(
-        RegExp(r'(\S+?|\[.+?|\{.+?)\.contains\((.+?)\)'),
-        (m) => '${m[2]} in ${m[1]}',
-      )
-      // Convert `range` to `x[i:j]
-      .replaceAllMapped(
-        RegExp(r'(\S+?)\.range\((.+?), (.+?)\)'),
-        (m) => '${m[1]}[${m[2]}:${m[3]}]',
-      );
-
-  final pass6 = pass5
-      // bool parsing
-      .replaceAllMapped(RegExp(r'parseBool\((.+?)\)'), (m) => 'bool(${m[1]})')
-      // bytes parsing
-      .replaceAllMapped(RegExp(r"parseBytes\('(.+?)'\)"), (m) => "b'${m[1]}'")
-      // float parsing
-      .replaceAllMapped(RegExp(r'parseFloat\((.+?)\)'), (m) => 'float(${m[1]})')
-      // int parsing
-      .replaceAllMapped(RegExp(r'parseInt\((.+?)\)'), (m) => 'int(${m[1]})');
-
-  final pass7 = pass6
-      // Raw rules string
-      .replaceAllMapped(RegExp(r'''raw\(['"](.+?)['"]\)'''), (m) => m[1]!);
-
-  final pass8 = translateEnums(pass7, {
-    'RulesDurationUnit': RulesDurationUnit.values,
-    'RulesMethod': RulesMethod.values,
-    'RulesIdentityProvider': RulesIdentityProvider.values,
-    'RulesSignInProvider': RulesSignInProvider.values,
+/// Extract raw rules strings, replace them with placeholders, sanitize the
+/// input, then replace the placeholders with the raw rules strings
+String transformIgnoringRaws(
+  String input,
+  List<String Function(String input)> transforms,
+) {
+  final raws = <String>[];
+  input =
+      input.replaceAllMapped(RegExp(r'''rules\.raw\(['"](.+?)['"]\)'''), (m) {
+    raws.add(m[1]!);
+    return '{RulesRawPlaceholder${raws.length - 1}}';
   });
 
-  final pass9 = translateAuthVariables(pass8);
+  input = transform(input, transforms);
 
-  final pass10 = pass9.replaceAll(
-    'resource.firestoreResourceName',
-    "resource['__name__']",
-  );
+  for (var i = 0; i < raws.length; i++) {
+    input = input.replaceFirst('{RulesRawPlaceholder$i}', raws.elementAt(i));
+  }
 
-  return pass10;
+  return input;
 }
 
 /// Sanitize path parameter prefixes from rules