+++ firebase_rules_generator/test_project/lib/firebase/raw/input.dart
@@ -5,6 +5,8 @@ import 'package:firebase_rules/firebase.dart';
 bool test() {
   final a = rules.raw("foo.bar.baz == 'qux'");
   final b = rules.raw('foo.bar.baz == 123');
+  // Make sure raws are completely ignored during sanitization
+  final c = rules.raw("foo.bar.rules == 'asdf'");
   return true;
 }