+++ firebase_rules_generator/CHANGELOG.md
@@ -1,5 +1,6 @@
 ## NEXT
 - Updates to support `firebase_rules` changes
+- Translate `r"..."` to `"..."` instead of `'...'`
 
 ## 0.0.1
 - Initial release