+++ firebase_rules_generator/CHANGELOG.md
@@ -1,6 +1,8 @@
 ## NEXT
 - Updates to support `firebase_rules` changes
 - Translate `r"..."` to `"..."` instead of `'...'`
+- Strip null check (`!`) operators
+- Ignore paths that equal `_`
 
 ## 0.0.1
 - Initial release