+++ firebase_rules_generator/CHANGELOG.md
@@ -4,6 +4,7 @@
 - Strip null check (`!`) operators
 - Ignore paths that equal `_`
 - Database string `matches(regex)` regex now must include the leading and trailing `/`
+- Fixes issue with raw matching
 
 ## 0.0.1
 - Initial release