+++ firebase_rules_generator/CHANGELOG.md
@@ -1,4 +1,4 @@
-## NEXT
+## 0.1.0
 - Updates to support `firebase_rules` changes
 - Translate `r"..."` to `"..."` instead of `'...'`
 - Strip null check (`!`) operators